fix portability concerns with size_t printfs
Apparently size_t is a concern for printf, as %lu may not be the correct size
to printf as on all platforms. There are two options to address this:
 - cast size_t to unsigned long
 - use %zu

This commit uses the former option, as the second option is apparently also
nonstandard for compilers, and also is dependent on C99. Casting seems to be
the "most" portable option, other than not using size_t, printf, or probably C.

+++ src/pseries.c
@@ -142,7 +142,8 @@ convert_to_ucs2 (const update_req_t *update_req, uint16_t **name)
 {
   if (update_req->label_size % 2 != 0)
     {
-      prlog (PR_ERR, "label has an odd number of bytes: %lu\n", update_req->label_size);
+      prlog (PR_ERR, "label has an odd number of bytes: %lu\n",
+             (unsigned long) update_req->label_size);
       return SV_LABEL_IS_NOT_WIDE_CHARACTERS;
     }
